+++ tools/python_tests/mbed_tools/build_tests/__init__.py
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2024 Jamie Smith. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+# Note: This folder cannot be named "build" to match the package because
+# pytest ignores any directories named "build" when discovering tests. Yes, really.
+# https://github.com/pytest-dev/pytest/issues/12625